+++ holy-water/StringExtention.cs
@@ -1,16 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace holy_water
 {
     public static class StringExtention
     {
-        public static double ToDouble(this string str)
+        public static decimal ToDecimal(this string str)
         {
-            return Convert.ToDouble(str);
+            return Convert.ToDecimal(str);
         }
     }
 }
